qt fix
